Move almost everything into the legacy directory.
