+++ src/legacy/components/pages/annual-report-2018/AnnualReportView.scss
@@ -1,4 +1,4 @@
-@import '../../../../stylesheets/utils/all';
+@import '../../../stylesheets/utils/all';
 
 .annual-report {
   align-items: center;
