+++ src/legacy/components/pages/donate/DonateView.scss
@@ -1,4 +1,4 @@
-@import '../../../../stylesheets/utils/all';
+@import '../../../stylesheets/utils/all';
 
 .donate-form-button {
   height: 44px;
