+++ src/legacy/images/favicon.ico
