+++ src/pages/annual-report-2018.js
@@ -1,6 +1,6 @@
 import React from "react";
 import AnnualReportView from "../legacy/components/pages/annual-report-2018/AnnualReportView";
-import Layout from "../layouts/Layout";
+import Layout from "../legacy/layouts/Layout";
 
 export default () => (
   <Layout>
