+++ src/pages/donate.js
@@ -1,6 +1,6 @@
 import React from "react";
 import DonateView from "../legacy/components/pages/donate/DonateView";
-import Layout from "../layouts/Layout";
+import Layout from "../legacy/layouts/Layout";
 
 export default () => (
   <Layout>
