+++ src/pages/events.js
@@ -1,6 +1,6 @@
 import React from "react";
 import EventsView from "../legacy/components/pages/events/EventsView";
-import Layout from "../layouts/Layout";
+import Layout from "../legacy/layouts/Layout";
 
 // Sync route definition
 export default () => (
