+++ src/pages/get-involved.js
@@ -1,6 +1,6 @@
 import React from "react";
 import GetInvolvedView from "../legacy/components/pages/get-involved/GetInvolvedView";
-import Layout from "../layouts/Layout";
+import Layout from "../legacy/layouts/Layout";
 
 // Sync route definition
 export default () => (
