+++ src/pages/givingtuesday.js
@@ -1,6 +1,6 @@
 import React from "react";
 import GivingTuesdayView from "../legacy/components/pages/givingtuesday/GivingTuesdayView";
-import Layout from "../layouts/Layout";
+import Layout from "../legacy/layouts/Layout";
 
 export default () => (
   <Layout>
