+++ src/pages/team.js
@@ -1,6 +1,6 @@
 import React from "react";
 import TeamView from "../legacy/components/pages/team/TeamView";
-import Layout from "../layouts/Layout";
+import Layout from "../legacy/layouts/Layout";
 
 export default () => (
   <Layout>
